requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
google-generativeai>=0.7.0
anyio
pytest-asyncio